Add progress bars for individual downloads
Additionally, this commit prevents the program from exiting if a single
resource has an error. Instead, that error will be noted (and details
will be given if the --verbose flag is passed), and downloading of
subsequent resources will continue.

One bug is that the `gdown` library manages its own progress bars, so
they do not appear in the same way as the ones generated here. There are
three possible solutions I see, from easiest to hardest

  1. Ignore it

  2. Pass the quiet option to gdown (eliminating any progress bar)

  3. Send a patch to gdown to accept a file descriptor to which to
     download, rather than a filename. (Unlikely to be accepted)

Taking (1) for now!