Add DownloadOptions model and --progress and --fail-on-error flags
This commit adds a new DownloadOptions model that encapsulates all of
the boolean flags that will affect how a download is run:

  1. `snippet_only` (default: False)
  2. `ignore_cache` (default: False)
  3. `verbose` (default: False)
  4. `progress` (default: False)
  5. `fail_on_error` (default: True)

The final two are new-- they control whether to show individual download
progress, and whether to abort the program when a download fails,
respectively.

This also fixes a bug where `snippet_only` was never passed from the
command line.